Docs & clean-up
