+++ packages/shortest/src/utils/errors.ts
@@ -1,3 +1,5 @@
+import { ZodError } from "zod";
+
 export type ConfigErrorType =
   | "duplicate-config"
   | "file-not-found"
@@ -41,3 +43,18 @@ export function getErrorDetails(error: any) {
         : undefined,
   };
 }
+
+export const formatZodError = <T>(
+  error: ZodError<T>,
+  label: string,
+): string => {
+  const errorsString = error.errors
+    .map((err) => {
+      const path = err.path.join(".");
+      const prefix = path ? `${path}: ` : "";
+      return `${prefix}${err.message}`;
+    })
+    .join("\n");
+
+  return `${label}\n${errorsString}`;
+};
